refactor: 각 컴포넌트의 스타일 통합
- AppScreen
- AppLayout
- FixedBottomCTA
