--- src/components/@common/AppLayout/AppLayout.styles.tsx
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Layout = styled.div`
-  max-width: 48rem;
-  min-height: calc((var(--vh, 1vh) * 100));
-  margin: 0 auto;
-`;
