+++ src/components/@common/AppScreen/components/BackArrow/index.tsx
@@ -1,0 +1,1 @@
+export * from './BackArrow';
